ENH: Significantly faster moment function - correcting sloppy mistakes

+++ scipy/stats/stats.py
@@ -945,7 +945,7 @@ def moment(a, moment=1, axis=0):
             return np.float64(0.0)
     else:
         # Subtract mean along the axis, compute element-wise square
-        a_zero_mean = np.float64(a - np.expand_dims(np.mean(a, axis), axis))
+        a_zero_mean = a - np.expand_dims(np.mean(a, axis), axis)
         a_zero_mean_2 = a_zero_mean**2
         
         s = a_zero_mean_2.copy()
